Moved controllers to file in backend, app not broken
