+++ server/server.js
@@ -3,15 +3,15 @@ const path = require('path');
 const express = require('express');
 const axios = require('axios');
 const coreJsCompat = require('@babel/preset-env/data/core-js-compat');
-const Controller = require('./controller');
-require("dotenv").config();
+const Controller = require('./controllers/controller');
+
 const cors = require('cors');
 const mongoose = require('mongoose');
 const cookieParser = require('cookie-parser');
 const session = require('express-session');
-const userController = require('./userController');
-const cookieController = require('./cookieController');
-const sessionController = require('./sessionController');
+const userController = require('./controllers/userController');
+const cookieController = require('./controllers/cookieController');
+const sessionController = require('./controllers/sessionController');
 
 const PORT = 3000;
 
@@ -22,13 +22,13 @@ app.use(express.urlencoded({ extended: true }));
 
 
 
-mongoose.connect(`mongodb+srv://${process.env.MONGOUSER}:${process.env.MONGOPASSWORD}@cluster0.txufs6f.mongodb.net/?retryWrites=true&w=majority`, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-    dbName: 'weatherApp'
-})
-    .then(() => console.log('Connected to Mongo DB.'))
-    .catch(err => console.log(err));
+// mongoose.connect(`mongodb+srv://${process.env.MONGOUSER}:${process.env.MONGOPASSWORD}@cluster0.txufs6f.mongodb.net/?retryWrites=true&w=majority`, {
+//     useNewUrlParser: true,
+//     useUnifiedTopology: true,
+//     dbName: 'weatherApp'
+// })
+//     .then(() => console.log('Connected to Mongo DB.'))
+//     .catch(err => console.log(err));
 
 app.get('/', (req, res) => {
     res.status(200).sendFile(path.join(__dirname, '../index.html'))
